move lists to helper file
